Move deserializer to its own file to avoid import issues, fix tests
